dashboard/requirements.txt
streamlit>=1.30
psycopg2-binary
pandas
plotly
requests
